requirements.txt
numpy
scikit-image
Pillow
scipy
matplotlib